requirements.txt
# متطلبات الأساسية
paho-mqtt==1.6.1
influxdb==5.3.2
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0

# متطلبات المحاكاة
faker==18.3.1
python-dotenv==1.1.1

# متطلبات الويب
Flask==2.2.3
Flask-Cors==3.0.10
requests==2.28.2
gunicorn==20.1.0

# متطلبات التصور
matplotlib==3.7.1
chart.js==2.9.4

# متطلبات التوثيق
sphinx==6.1.3
sphinx-rtd-theme==1.2.0

# متطلبات الاختبار
pytest==7.3.1
pytest-mock==3.10.0

pydantic==2.11.7
schedule==1.2.2